social_media_platform_project/requirements/base.txt
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
 drf-spectacular-sidecar==2024.7.1 
gunicorn==23.0.0
django-filter==24.3 

